feat(suite): create useMessagingSystem hook
It may be used in multiple places like: SuiteLayout on Onboarding
